--- apps/gym/src/app/Home.module.css
@@ -1,38 +1,0 @@
-.container {
-  text-align: center;
-  margin-top: 40px;
-  font-size: 150%;
-}
-
-.container h1 {
-  margin: 0;
-}
-
-.description {
-  width: 50%;
-  margin: 40px auto 0px auto;
-}
-
-.thanks {
-  margin: 40px auto;
-}
-
-.link {
-  display: block;
-  width: 100%;
-  height: 100%;
-}
-
-.buttons {
-  display: flex;
-  flex-direction: column;
-  margin: auto;
-  width: max-content;
-  gap: 10px;
-}
-
-@media (width < 480px) {
-  .description {
-    width: 90%;
-  }
-}
