--- apps/gym/src/components/isolated/Login/Credentials/Credentials.module.css
@@ -1,24 +1,0 @@
-.container {
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-  padding: 16px;
-  border-radius: 8px;
-  box-shadow: var(--shadow-primary);
-  width: max-content;
-  max-width: calc(100% - 96px);
-  position: absolute;
-  bottom: 32px;
-  right: 32px;
-}
-
-.container h3 {
-  margin-bottom: 8px;
-}
-
-@media (width < 600px) {
-  .container {
-    position: unset;
-    max-width: calc(100% - 32px);
-  }
-}
