--- apps/gym/src/components/reusable/Header/Header.module.css
@@ -1,4 +1,0 @@
-.container {
-  width: 100%;
-  border-bottom: 1px solid var(--color-light-gray);
-}
