--- apps/gym/src/components/reusable/Header/HeaderDesktop/HeaderDesktop.module.css
@@ -1,21 +1,0 @@
-.container {
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  padding: 10px 0;
-  width: 95%;
-  margin: auto;
-  font-size: 14px;
-  gap: 40px;
-}
-
-.label {
-  font-size: 12px;
-  color: var(--color-dark-gray);
-}
-
-@media (width < 480px) {
-  .container {
-    display: none;
-  }
-}
