--- apps/gym/src/components/reusable/Header/HeaderMobile/DropdownMobile/DropdownMobile.module.css
@@ -1,29 +1,0 @@
-.container {
-  display: flex;
-  flex-direction: column;
-  position: absolute;
-  inset: 0;
-  background-color: var(--color-primary);
-  width: calc(100vw - 16px);
-  height: calc(100vh - 16px);
-  overflow: hidden;
-  padding: 8px;
-  gap: 4px;
-  z-index: 1000000;
-}
-
-.option {
-  font-size: 3rem;
-  /* font-weight: 500; */
-  color: white;
-}
-
-.logout {
-  cursor: pointer;
-  position: absolute;
-  bottom: 16px;
-  display: flex;
-  align-items: center;
-  gap: 16px;
-  margin-left: 8px;
-}
