--- apps/gym/src/components/reusable/Header/HeaderMobile/HeaderMobile.module.css
@@ -1,12 +1,0 @@
-.container {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 0 8px;
-}
-
-@media (width > 480px) {
-  .container {
-    display: none;
-  }
-}
