+++ apps/gym/src/components/reusable/Header/header-mobile.tsx
@@ -1,11 +1,10 @@
 'use client'
 
-import styles from './HeaderMobile.module.css'
 import { HeaderMobileProps } from '@/types'
 import { Hamburger } from '@gymapp/gymui/Hamburger'
 import { useToggle } from '@/hooks'
-import { DropdownMobile } from './DropdownMobile/DropdownMobile'
-import { DesktopLogo } from '../../Logo/DesktopLogo'
+import { DropdownMobile } from './dropdown-mobile'
+import { DesktopLogo } from '../Logo/DesktopLogo'
 import { AnimatePresence } from 'motion/react'
 
 export const HeaderMobile = ({}: HeaderMobileProps) => {
@@ -16,7 +15,7 @@ export const HeaderMobile = ({}: HeaderMobileProps) => {
   }
 
   return (
-    <div className={styles.container}>
+    <div className='flex items-center justify-between px-2 md:hidden'>
       <DesktopLogo />
       <div>
         <Hamburger active={active} onClick={toggleDropdown} />
