+++ apps/gym/src/components/reusable/Header/index.ts
@@ -1,1 +1,1 @@
-export * from './Header'
+export * from './header'
